feat(web): move components
